% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_helper_funcs.R
\name{appendDetections}
\alias{appendDetections}
\title{Internal function to append needed columns to table detections}
\usage{
appendDetections(sync_dat)
}
\arguments{
\item{sync_dat}{List containing data.tables with hydrophone information and detections.}
}
\description{
Internal function to append needed columns to table detections
}
