% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_helper_funcs.R
\name{buildToaListGross}
\alias{buildToaListGross}
\title{Internal function to build gross TOA matrix, sync_tag_vec and epo_self_vec - the matrix needs pruning before use in sync function}
\usage{
buildToaListGross(sync_dat)
}
\arguments{
\item{sync_dat}{List containing data.tables with hydrophone information and detections.}
}
\description{
Internal function to build gross TOA matrix, sync_tag_vec and epo_self_vec - the matrix needs pruning before use in sync function
}
