% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInp.R
\name{checkInp}
\alias{checkInp}
\title{Check consistency of `inp` object obtained from `getInp()`}
\usage{
checkInp(inp)
}
\arguments{
\item{inp}{Object obtained using `getInp()`}
}
\description{
Check consistency of `inp` object obtained from `getInp()`
}
