% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_checkInpSyncData.R
\name{checkInpSyncData}
\alias{checkInpSyncData}
\title{Checks input data for sync models...}
\usage{
checkInpSyncData(hydros, dat_sync, dat_ss, sync_params)
}
\description{
Checks input data for sync models...
}
