% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_align}
\alias{dat_align}
\title{Example data for showcasing yaps function alignBurstSeq()}
\format{
A list containing 2 items:
\describe{
  \item{synced_dat_1315}{
     data.table containing synced detections of tag 1315.
  }
  \item{synced_dat_1315}{
     vector of small part of the complete sequence of known random BIs.
  }
}
}
\usage{
dat_align
}
\description{
Function alignBurstSeq() is used to align synced detection data with a sequence of known random burst intervals (BI). \cr
This step is needed to take advantage of the extra information available when working with random BI data with a known sequence. \cr
This small sample is obtained from the accompanying data package `yapsdata`.
}
\keyword{datasets}
