% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTrack.R
\name{extractTrack}
\alias{extractTrack}
\title{Internal function to extract the estimated track}
\usage{
extractTrack(inp, pl, plsd)
}
\arguments{
\item{inp}{input}

\item{pl}{input}

\item{plsd}{input}
}
\description{
Internal function to extract the estimated track
}
