% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fineTuneSyncModel.R
\name{fineTuneSyncModel}
\alias{fineTuneSyncModel}
\title{Fine-tune an already fitted sync_model
Wrapper function to re-run getSyncModel() using the same data, but excluding outliers. Note dimensions of data might change if eps_threshold results in empty rows in the TOA-matrix.}
\usage{
fineTuneSyncModel(sync_model, eps_threshold, silent = TRUE)
}
\arguments{
\item{sync_model}{sync_model obtained using getSyncModel()}

\item{eps_threshold}{Maximum value of residual measured in meter assuming speed of sound = 1450 m/s}

\item{silent}{logical whether to make getSyncModel() silent}
}
\description{
Fine-tune an already fitted sync_model
Wrapper function to re-run getSyncModel() using the same data, but excluding outliers. Note dimensions of data might change if eps_threshold results in empty rows in the TOA-matrix.
}
