% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_getters.R
\name{getDatTmbSync}
\alias{getDatTmbSync}
\title{Internal function to get dat for TMB sync}
\usage{
getDatTmbSync(inp_H_info, time_keeper_idx, inp_toa_list, fixed_hydros_vec,
  offset_vals, ss_vals, T0)
}
\description{
Internal function to get dat for TMB sync
}
