% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_getters.R
\name{getFixedHydrosVec}
\alias{getFixedHydrosVec}
\title{Internal function to get vector of which hydros are fixed}
\usage{
getFixedHydrosVec(sync_dat, fixed_hydros_idx)
}
\description{
Internal function to get vector of which hydros are fixed
}
