% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepTmb.R
\name{getInpParams}
\alias{getInpParams}
\title{Get parameters for this specific data set}
\usage{
getInpParams(hydros, toa, pingType)
}
\arguments{
\item{hydros}{Dataframe from simHydros() or Dataframe with columns hx and hy containing positions of the receivers. Translate the coordinates to get the grid centre close to (0;0).}

\item{toa}{TOA-matrix: matrix with receivers in rows and detections in columns. Make sure that the receivers are in the same order as in hydros, and that the matrix is very regular: one ping per column (inlude empty columns if a ping is not detected).}

\item{pingType}{Type of transmitter to simulate - either stable burst interval ('sbi'), random burst interval ('rbi') or random burst interval but where the random sequence is known a priori}
}
\description{
Compile a list of relevant parameters (e.g. T0) to use later on
}
