% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncHelperFuncs.R
\name{getInpSyncHInfo}
\alias{getInpSyncHInfo}
\title{Internal function. Get hydros for sync from sync_dat}
\usage{
getInpSyncHInfo(sync_dat)
}
\arguments{
\item{sync_dat}{List containing data.tables with hydrophone information and detections. See e.g. `?ssu1` for example}
}
\description{
Internal function. Get hydros for sync from sync_dat
}
