% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_getters.R
\name{getInpSyncToaList}
\alias{getInpSyncToaList}
\title{Internal function. Get toa for sync from sync_dat}
\usage{
getInpSyncToaList(sync_dat, max_epo_diff, min_hydros)
}
\arguments{
\item{sync_dat}{List containing data.tables with hydrophone information and detections.}
}
\description{
Internal function. Get toa for sync from sync_dat
}
