% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_getLinCorr.R
\name{getLinCorr}
\alias{getLinCorr}
\title{Get a matrix of coefficients for linear models used for linear correction prior to run getSyncModel()
To be used as input in getInpSync() parameter lin_corr_coeffs}
\usage{
getLinCorr(sync_model)
}
\description{
Get a matrix of coefficients for linear models used for linear correction prior to run getSyncModel()
To be used as input in getInpSync() parameter lin_corr_coeffs
}
