% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_getters.R
\name{getOffsetVals}
\alias{getOffsetVals}
\title{Internal function to get info relating to sync offsets per day}
\usage{
getOffsetVals(inp_toa_list, n_offset_day)
}
\description{
Internal function to get info relating to sync offsets per day
}
