% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepTmb.R
\name{getParams}
\alias{getParams}
\title{Get params-list for use in TMB}
\usage{
getParams(datTmb)
}
\arguments{
\item{datTmb}{Object obtained using getDatTmb()}
}
\value{
List of params for use in TMB
}
\description{
Compile a list of parameters for use in TMB.
}
