% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_getSmoothOffsets.R
\name{getSmoothOffsets}
\alias{getSmoothOffsets}
\title{Get smoothed sync offsets from sync_model}
\usage{
getSmoothOffsets(sync_model, E_smooth = "gaus", k_div = 2)
}
\description{
Get smoothed sync offsets from sync_model
}
