% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_getters.R
\name{getSsVals}
\alias{getSsVals}
\title{Internal function to get info relating to sync ss per day}
\usage{
getSsVals(inp_toa_list, n_ss_day)
}
\description{
Internal function to get info relating to sync ss per day
}
