% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_getSyncCheckDat.R
\name{getSyncCheckDat}
\alias{getSyncCheckDat}
\title{Get data for check the sync model from sync_model}
\usage{
getSyncCheckDat(sync_model)
}
\description{
Get data for check the sync model from sync_model
}
