% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInpSync.R
\name{getSyncCoverage}
\alias{getSyncCoverage}
\title{Quick overview to check if all hydros have enough data within each offset period.}
\usage{
getSyncCoverage(inp_sync, plot = FALSE)
}
\description{
Quick overview to check if all hydros have enough data within each offset period.
}
