% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getToaYaps.R
\name{getToaYaps}
\alias{getToaYaps}
\title{Build TOA matrix from synced data.table - also do some pre-filtering of severe MP, pruning loose ends etc}
\usage{
getToaYaps(synced_dat, hydros, rbi_min, rbi_max)
}
\arguments{
\item{synced_dat}{`data.table` containing synchronized data formatted as output from/or obtained using `applySync()`}

\item{hydros}{Dataframe from simHydros() or Dataframe with columns hx and hy containing positions of the receivers. Translate the coordinates to get the grid centre close to (0;0).}

\item{rbi_min}{Minimum and maximum BI for random burst interval transmitters}

\item{rbi_max}{Minimum and maximum BI for random burst interval transmitters}
}
\description{
Build TOA matrix from synced data.table - also do some pre-filtering of severe MP, pruning loose ends etc
}
