% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBbox.R
\name{plotBbox}
\alias{plotBbox}
\title{Graphical representation of spatial constraints}
\usage{
plotBbox(hydros, bbox)
}
\arguments{
\item{hydros}{Dataframe from simHydros() or Dataframe with columns hx and hy containing positions of the receivers. Translate the coordinates to get the grid centre close to (0;0).}

\item{bbox}{Spatial constraints in the form of a bounding box. See ?getBbox for details.}
}
\description{
Graphical representation of spatial constraints
}
\examples{
hydros <- ssu1$hydros
colnames(hydros) <- c('serial','hx','hy','hz','sync_tag','idx')
bbox <- getBbox(hydros)
plotBbox(hydros, bbox)
}
