% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncPlotters.R
\name{plotSyncModelHydros}
\alias{plotSyncModelHydros}
\title{Plot hydrophone positions. Especially useful if some hydro re-positioned as part of the sync model.}
\usage{
plotSyncModelHydros(sync_model)
}
\arguments{
\item{sync_model}{Synchronization model obtained using `getSyncModel()`}
}
\description{
Plot hydrophone positions. Especially useful if some hydro re-positioned as part of the sync model.
}
