% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotYaps.R
\name{plotYaps}
\alias{plotYaps}
\title{Basic plots of yaps output}
\usage{
plotYaps(yaps_out, type = "map", xlim = NULL, ylim = NULL, main = NULL)
}
\arguments{
\item{yaps_out}{Output from succesful run of \code{runYaps()}}

\item{type}{Plot type. \code{type="map"} prodces a basic map of estimated track and hydrophones; \code{type="coord_X"}, \code{type="coord_Y"} produces plots of X and Y coordinated including +- 1 standard error.}

\item{xlim, ylim}{Optional vectors of length 2 to set xlim and/or ylim.}

\item{main}{Title of plot - optional.}
}
\value{
No return value, called to plot graphics.
}
\description{
Basic plots of yaps output
}
\examples{
library(yaps)
plotYaps(ssu1$yaps_out, type="map")
plotYaps(ssu1$yaps_out, type="coord_X")
plotYaps(ssu1$yaps_out, type="coord_Y")

}
