% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepFiles.R
\name{prepDetections}
\alias{prepDetections}
\title{Experimental! Prepare detections data.table from raw data - csv-files exported from vendor software}
\usage{
prepDetections(raw_dat, type)
}
\arguments{
\item{raw_dat}{Data file from vendor supplied software}

\item{type}{Type of the vendor file. Currently only 'vemco_vue' is supported.}
}
\value{
\code{data.table} containing detections extracted from manufacturer data file.
}
\description{
Experimental! Prepare detections data.table from raw data - csv-files exported from vendor software
}
\examples{
\dontrun{
prepped_detections <- prepDetections("path-to-raw-data-file", type="vemco_vue")
}
}
