% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_helper_funcs.R
\name{pruneToaListGross}
\alias{pruneToaListGross}
\title{Internal function to prune the object obtained by buildToaListGross}
\usage{
pruneToaListGross(toa_list_gross, max_epo_diff, min_hydros)
}
\description{
Internal function to prune the object obtained by buildToaListGross
}
