% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeMultipath.R
\name{removeMultipath}
\alias{removeMultipath}
\title{Remove multipath detections from detection table}
\usage{
removeMultipath(dets, mp_threshold = 0.5)
}
\description{
Remove multipath detections from detection table
}
