% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTrack.R
\name{simHydros}
\alias{simHydros}
\title{Sim hydrophone array configuration}
\usage{
simHydros(auto = TRUE, trueTrack = NULL)
}
\arguments{
\item{auto}{If TRUE, attempts to find a decent array configuration to cover the simulated true track.}

\item{trueTrack}{Track obtained from simTrueTrack().}
}
\value{
Dataframe containing X and Y for hydros
}
\description{
Sim hydrophone array configuration
}
