% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTrack.R
\name{simToa}
\alias{simToa}
\title{Sim TOA matrix for the supplied telemetryTrack}
\usage{
simToa(telemetryTrack, hydros, pingType, sigmaToa, pNA, pMP)
}
\arguments{
\item{telemetryTrack}{Dataframe obtained from simTelemetryTrack}

\item{hydros}{Dataframe obtained from getHydros}

\item{pingType}{Type of transmitter to simulate - either stable burst interval ('sbi'), random burst interval ('rbi') or random burst interval but where the random sequence is known a priori}

\item{sigmaToa}{Detection uncertainty}

\item{pNA}{Probability of missing detection 0-1}

\item{pMP}{Probability of multipath propagated signal 0-1}
}
\value{
List containing TOA matrix (toa) and matrix indicating, which obs are multipath (mp_mat)
}
\description{
Provides the TOA matrix for the specified telemetryTrack. Probability of NA (pNA) and observation noise (sigmaToa) can be specified.
}
