% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTrack.R
\name{simTOP}
\alias{simTOP}
\title{Simulate time of pings (Internal function)}
\usage{
simTOP(trueTrack, pingType, sbi_mean = NULL, sbi_sd = NULL,
  rbi_min = NULL, rbi_max = NULL)
}
\arguments{
\item{trueTrack}{Track obtained from simTrueTrack().}

\item{pingType}{Type of transmitter to simulate - either stable burst interval (sbi) or random burst interval (rbi)}

\item{sbi_mean}{Mean and SD of burst interval when pingType = 'sbi'}

\item{sbi_sd}{Mean and SD of burst interval when pingType = 'sbi'}

\item{rbi_min}{Minimum and maximum BI for random burst interval transmitters}

\item{rbi_max}{Minimum and maximum BI for random burst interval transmitters}
}
\value{
Vector of simulated time of pings. Length = number of steps in simulated track.
}
\description{
This function will produce simulated time of pings of the relevant type. For internal use only; should not be needed by users.
}
