% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testYaps.R
\name{testYaps}
\alias{testYaps}
\title{Test YAPS core functionality}
\usage{
testYaps(
  silent = TRUE,
  pingType = "sbi",
  est_ss = TRUE,
  opt_fun = "nlminb",
  opt_controls = list(),
  return_yaps = FALSE,
  tmb_smartsearch = TRUE
)
}
\arguments{
\item{silent}{Logical whether to print output to the console}

\item{pingType}{Type of transmitter to simulate - either stable burst interval ('sbi'), random burst interval ('rbi') or random burst interval but where the random sequence is known a priori}

\item{est_ss}{Logical whether to test using ss_data_what = 'est' (est_ss = TRUE) or ss_data_what = 'data' (est_ss = FALSE)}

\item{return_yaps}{Logical whether to return the fitted yaps model. Default=FALSE.}
}
\value{
If \code{return_yaps == TRUE}, the fitted \code{yaps} object. See \code{?runYaps} for further info.
}
\description{
Run \code{testYaps()} to check that the core functions of YAPS is working correctly.
Output should be a random simulated (black) and estimated (red) track.
}
\examples{
\donttest{
#' # To test  basic functionality of yaps using simulated data
testYaps()
# # # Three pingTypes are availabe: 
# # #	fixed burst interval (testYaps(pingType='sbi')), 
# # # 	random burst interval with UNKNOWN burst interval sequence('testYaps(pingType='rbi')), 
# # # 	random burst interval with KNOWN burst interval sequence (testYaps(pingType='pbi'))
}
}
