% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yapsify.R
\name{yapsify}
\alias{yapsify}
\title{Internal function. yapsify / standarize hydros and detections for sync and yaps}
\usage{
yapsify(yapsify_me)
}
\description{
Internal function. yapsify / standarize hydros and detections for sync and yaps
}
