% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncHelperFuncs.R
\name{appendDetections}
\alias{appendDetections}
\title{Internal function to append needed columns to table detections}
\usage{
appendDetections(sync_dat)
}
\arguments{
\item{sync_dat}{List containing data.tables with hydrophone information and detections. See e.g. `?ssu1` for example}
}
\description{
Internal function to append needed columns to table detections
}
