% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncHelperFuncs.R
\name{buildToaListGross}
\alias{buildToaListGross}
\title{Internal function to build gross TOA matrix, sync_tag_vec and epo_self_vec - the matrix needs pruning before use in sync function}
\usage{
buildToaListGross(sync_dat, excl_self_detect)
}
\arguments{
\item{sync_dat}{List containing data.tables with hydrophone information and detections. See e.g. `?ssu1` for example}

\item{excl_self_detect}{Logical whether to excluded detections of sync tags on the hydros they are co-located with. Sometimes self detections can introduce excessive residuals in the sync model in which case they should be excluded.}
}
\description{
Internal function to build gross TOA matrix, sync_tag_vec and epo_self_vec - the matrix needs pruning before use in sync function
}
