% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepTmb.R
\name{getDatTmb}
\alias{getDatTmb}
\title{Get data for input to TMB}
\usage{
getDatTmb(hydros, toa, E_dist, n_ss, pingType, rbi_min, rbi_max)
}
\arguments{
\item{hydros}{Data frame from simHydros()}

\item{toa}{TOA-matrix}

\item{E_dist}{Which distribution to use in the model - 0=Gaussian, 1=Mixture of Gauss and t}

\item{n_ss}{Number of speed of sound to estimate}

\item{pingType}{Type of transmitter to simulate - either stable burst interval (sbi) or random burst interval (rbi)}

\item{rbi_min}{Minimum and maximum BI for random burst interval transmitters}

\item{rbi_max}{Minimum and maximum BI for random burst interval transmitters}
}
\value{
List for use in TMB.
}
\description{
Compile data for input to TMB.
}
