% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepTmb.R
\name{getInp}
\alias{getInp}
\title{Get prepared inp-object for use in TMB-call}
\usage{
getInp(hydros, toa, E_dist, n_ss, pingType, sdInits = 1, rbi_min = 0,
  rbi_max = 0)
}
\arguments{
\item{hydros}{Data frame from simHydros()}

\item{toa}{TOA-matrix}

\item{E_dist}{Which distribution to use in the model - 0=Gaussian, 1=Mixture of Gauss and t}

\item{n_ss}{Number of speed of sound to estimate}

\item{pingType}{Type of transmitter to simulate - either stable burst interval (sbi) or random burst interval (rbi)}

\item{sdInits}{If >0 initial values will be randomized around the normally fixed value using rnorm(length(inits), mean=inits, sd=sdInits)}

\item{rbi_min, rbi_max}{Minimum and maximum BI for random burst interval transmitters}
}
\value{
List of input data ready for use in TMB-call
}
\description{
Wrapper-function to compile a list of inoput needed to run TMB
}
