% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_getters.R
\name{getInpSync}
\alias{getInpSync}
\title{Get object inp for synchronization}
\usage{
getInpSync(sync_dat, max_epo_diff, min_hydros, time_keeper_idx,
  fixed_hydros_idx, n_offset_day, n_ss_day)
}
\arguments{
\item{sync_dat}{List containing data.tables with hydrophone information and detections.}
}
\description{
Get object inp for synchronization
}
