% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInpSync.R
\name{getSyncCoverage}
\alias{getSyncCoverage}
\title{Quick overview to check if all hydros have enough data within each offset period.}
\usage{
getSyncCoverage(inp_sync, plot = FALSE)
}
\arguments{
\item{inp_sync}{Object obtained using `getInpSync()`}

\item{plot}{Logical indicating whether to plot a visual or not.}
}
\description{
Quick overview to check if all hydros have enough data within each offset period.
}
