% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncGetters.R
\name{getSyncModel}
\alias{getSyncModel}
\title{Get sync model from inp_sync object obtained by getInpSync()}
\usage{
getSyncModel(inp_sync, silent = TRUE, fine_tune = TRUE)
}
\arguments{
\item{inp_sync}{Input data prepared for the sync model using `getInpSync()`}

\item{silent}{Keep TMB quiet}

\item{fine_tune}{Logical. Wheter to re-run the sync model excluding residual outliers}
}
\description{
Get sync model from inp_sync object obtained by getInpSync()
}
