% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getToaRbi.R
\name{getToaRbi}
\alias{getToaRbi}
\title{Get TOA matrix for random burst interval transmitters}
\usage{
getToaRbi(dets, hydros, rbi_min, rbi_max)
}
\description{
Get TOA matrix for random burst interval transmitters
}
