% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncPlotters.R
\name{plotSyncModelHydros}
\alias{plotSyncModelHydros}
\title{Plot hydrophone positions. Especially useful if some hydro re-positioned as part of the sync model.}
\usage{
plotSyncModelHydros(sync_model)
}
\arguments{
\item{sync_model}{Synchronization model obtained using `getSyncModel()`}

\item{by}{What to facet/group the plot by? Currently supports one of 'sync_bin_sync', 'sync_bin_hydro', 'sync_bin_sync_smooth', 'sync_bin_hydro_smooth', 'hydro', 'sync_tag'}
}
\description{
Plot hydrophone positions. Especially useful if some hydro re-positioned as part of the sync model.
}
