% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotYaps.R
\name{plotYaps}
\alias{plotYaps}
\title{Basic plots of yaps output}
\usage{
plotYaps(inp, yaps_out, type = "map")
}
\arguments{
\item{inp}{Input object obtained using \code{getInp()}}

\item{yaps_out}{Output from succesful run of \code{runYaps()}}

\item{type}{Plot type. \code{type="map"} prodces a basic map of estimated track and hydrophones; \code{type="coord_X"}, \code{type="coord_Y"} produces plots of X and Y coordinated including +- 1 standard error.}
}
\description{
Basic plots of yaps output
}
