% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runYaps.R
\name{runYaps}
\alias{runYaps}
\title{Wrapper function to doRunYaps()}
\usage{
runYaps(inp, silent, best_of = 1)
}
\arguments{
\item{inp}{input}

\item{silent}{input}
}
\description{
Wrapper function to doRunYaps()
}
