% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runYaps.R
\name{runYaps}
\alias{runYaps}
\alias{runTmb}
\title{Function to run TMB to estimate track}
\usage{
runYaps(inp, maxIter = 1000, getPlsd = TRUE, getRep = TRUE,
  silent = TRUE)

runTmb(inp, maxIter = 1000, getPlsd = TRUE, getRep = TRUE,
  silent = TRUE)
}
\arguments{
\item{inp}{inp-object obtained from getInp()}

\item{maxIter}{Sets inner.control(maxit) of the tmb-call. Increase if model is not converging.}

\item{getPlsd, getRep}{Whether or not to get sd estimates (plsd=TRUE) and reported values (getRep=TRUE).}

\item{silent}{Logical whether to keep the optimization quiet.}
}
\description{
Function to run TMB to estimate track
}
