% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testYaps.R
\name{testYaps}
\alias{testYaps}
\title{Test YAPS core functionality}
\usage{
testYaps()
}
\description{
Run `testYaps()` to check that the core functions of YAPS is working correctly.
Output should be a random simulated (black) and estimated (red) track.
}
