% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testYaps.R
\name{testYaps}
\alias{testYaps}
\title{Test YAPS core functionality}
\usage{
testYaps(silent = FALSE, pingType = "sbi", est_ss = TRUE)
}
\arguments{
\item{silent}{Logical whether to print output to the console}
}
\description{
Run `testYaps()` to check that the core functions of YAPS is working correctly.
Output should be a random simulated (black) and estimated (red) track.
}
